+++ src/main/java/br/com/expresscart/exception/handler/BusinessException.java
@@ -1,4 +1,4 @@
-package br.com.expresscart.exceptions;
+package br.com.expresscart.exception.handler;
 
 public class BusinessException extends RuntimeException {
