+++ src/main/java/br/com/expresscart/exception/handler/DataNotFoundException.java
@@ -1,4 +1,4 @@
-package br.com.expresscart.exceptions;
+package br.com.expresscart.exception.handler;
 
 public class DataNotFoundException extends RuntimeException {
     public DataNotFoundException(String message) {
